Add styling for Developer Tools
* Add instructions on readme to add dev tools styling
* Add `_dev-tools.css` with css variables

It seems that not all of the dev tools is themeable. Some of the colors
come from whether you are in light or dark theme. I was not able to get
the light theme to look good by adjusting the css variables. Many panels
had a white background with dark text. Because of this I made this
feature an opt-in thing so that users can see the instructions of
setting the theme to dark.